remoe model/string.go
