--- cdc/model/string.go
@@ -1,44 +1,0 @@
-// Copyright 2020 PingCAP, Inc.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package model
-
-import (
-	"strings"
-
-	cerror "github.com/pingcap/tiflow/pkg/errors"
-)
-
-// HolderString returns a string of place holders separated by comma
-// n must be greater or equal than 1, or the function will panic
-func HolderString(n int) string {
-	var builder strings.Builder
-	builder.Grow((n-1)*2 + 1)
-	for i := 0; i < n; i++ {
-		if i > 0 {
-			builder.WriteString(",")
-		}
-		builder.WriteString("?")
-	}
-	return builder.String()
-}
-
-// ExtractKeySuffix extracts the suffix of an etcd key, such as extracting
-// "6a6c6dd290bc8732" from /tidb/cdc/changefeed/config/6a6c6dd290bc8732
-func ExtractKeySuffix(key string) (string, error) {
-	subs := strings.Split(key, "/")
-	if len(subs) < 2 {
-		return "", cerror.ErrInvalidEtcdKey.GenWithStackByArgs(key)
-	}
-	return subs[len(subs)-1], nil
-}
