+++ cdc/sink/mysql/mysql.go
@@ -844,7 +844,7 @@ func prepareReplace(
 		builder.WriteString("REPLACE INTO " + quoteTable + colList + " VALUES ")
 	}
 	if appendPlaceHolder {
-		builder.WriteString("(" + model.HolderString(len(columnNames)) + ");")
+		builder.WriteString("(" + placeHolder(len(columnNames)) + ");")
 	}
 
 	return builder.String(), args
@@ -855,7 +855,7 @@ func prepareReplace(
 // args: (1,"",2,"2",3,"")
 func reduceReplace(replaces map[string][][]interface{}, batchSize int) ([]string, [][]interface{}) {
 	nextHolderString := func(query string, valueNum int, last bool) string {
-		query += "(" + model.HolderString(valueNum) + ")"
+		query += "(" + placeHolder(valueNum) + ")"
 		if !last {
 			query += ","
 		}
@@ -1018,4 +1018,18 @@ func getDBConn(ctx context.Context, dsnStr string) (*sql.DB, error) {
 		return nil, cerror.ErrMySQLConnectionError.Wrap(err).GenWithStack("fail to open MySQL connection")
 	}
 	return db, nil
+}
+
+// placeHolder returns a string separated by comma
+// n must be greater or equal than 1, or the function will panic
+func placeHolder(n int) string {
+	var builder strings.Builder
+	builder.Grow((n-1)*2 + 1)
+	for i := 0; i < n; i++ {
+		if i > 0 {
+			builder.WriteString(",")
+		}
+		builder.WriteString("?")
+	}
+	return builder.String()
 }
